remove url and music channel implemented

+++ src/index.ts
@@ -1,15 +1,16 @@
 import { Channel, Client, Guild, GuildMember, Intents, Message, MessageReaction, PartialDMChannel, PartialGuildMember, PartialMessage, PartialUser, Presence, StreamDispatcher, TextChannel, User, VoiceState } from "discord.js";
 import mongoose from 'mongoose'; // we want to load an object not only functions
 import command_config_json from './config.command.json';
+import event_config_json from './config.event.json';
 import config from './config.json';
 import { included_in_url_list } from './libraries/guildOps';
 import { is_authorised, is_url, message_reply, pad, time_elapsed } from './libraries/helpOps';
 import { client_talk } from './libraries/localisationOps';
 import { isProfane } from "./libraries/modOps";
-import { fetch_guild } from "./libraries/mongoOps";
-import { start } from './libraries/musicOps';
+import { fetch_guild, remove_url, set_music_data } from "./libraries/mongoOps";
+import { start, stop } from './libraries/musicOps';
 import { add_points_message } from './libraries/userOps';
-import { GuildPrtl } from './types/classes/GuildPrtl';
+import { GuildPrtl, MusicData } from './types/classes/GuildPrtl';
 import { ActiveCooldowns, CommandOptions, ReturnPormise } from "./types/interfaces/InterfacesPrtl";
 const AntiSpam = require('discord-anti-spam');
 
@@ -169,99 +170,116 @@ client.on('message', async (message: Message) => {
 	if (message.channel.type === 'dm') return;
 
 	// check if something written in portal channels
-	const handled = await portal_channel_handler(message);
-	if (handled) return;
+	portal_channel_handler(message)
+		.then(r => {
+			if (!r) {
+				// ranking system
+				ranking_system(message);
 
-	// ranking system
-	ranking_system(message);
+				anti_spam.message(message);
 
-	anti_spam.message(message);
+				// const spam = isSpam(message.content);
+				// if (spam.length > 0) {
+				// 	message.react('🚩');
+				// 	message.author
+				// 		.send(`try not to spam`)
+				// 		.catch(console.error);
+				// }
 
-	// const spam = isSpam(message.content);
-	// if (spam.length > 0) {
-	// 	message.react('🚩');
-	// 	message.author
-	// 		.send(`try not to spam`)
-	// 		.catch(console.error);
-	// }
+				// profanity check
+				const profanities = isProfane(message.content);
+				if (profanities.length > 0) {
+					message.react('🚩');
+					message.author
+						.send(`try not to use profanities (${profanities.join(',')})`)
+						.catch(console.error);
+				}
 
-	// profanity check
-	const profanities = isProfane(message.content);
-	if (profanities.length > 0) {
-		message.react('🚩');
-		message.author
-			.send(`try not to use profanities (${profanities.join(',')})`)
-			.catch(console.error);
-	}
+				// Ignore any message that does not start with prefix
+				if (message.content.indexOf(config.prefix) !== 0) return;
 
-	// Ignore any message that does not start with prefix
-	if (message.content.indexOf(config.prefix) !== 0) return;
+				// Separate function name, and arguments of function
+				const args = message.content.slice(config.prefix.length).trim().split(/ +/g);
 
-	// Separate function name, and arguments of function
-	const args = message.content.slice(config.prefix.length).trim().split(/ +/g);
+				const cmd_only = args.shift();
+				if (!cmd_only) return;
 
-	const cmd_only = args.shift();
-	if (!cmd_only) return;
+				const cmd = cmd_only.toLowerCase();
 
-	const cmd = cmd_only.toLowerCase();
+				let path_to_command: string = '';
+				let command_options: CommandOptions | undefined;
+				let type: string = 'none';
 
+				const is_portal_command = command_config_json.some(category => {
+					command_options = category.commands.find(command => command.name === cmd);
+					if (command_options) {
+						type = command_options.range;
+						path_to_command = category.path;
+						return true;
+					}
+					return false;
+				});
 
-	let path_to_command: string = '';
-	let command_options: CommandOptions | undefined;
-	let type: string = 'none';
-
-	const is_portal_command = command_config_json.some(category => {
-		command_options = category.commands.find(command => command.name === cmd);
-		if (command_options) {
-			type = command_options.range;
-			path_to_command = category.path;
-			return true;
-		}
-		return false;
-	});
-
-	// is not a portal command
-	if (!is_portal_command) return false;
-
-	if (!command_options) {
-		message_reply(false, message.channel, message, message.author,
-			'could not get command option', undefined, client);
-		return false;
-	}
-
-	fetch_guild(message.guild.id)
-		.then(guild_object => {
-			if (!guild_object) {
-				message_reply(false, message.channel, message, message.author,
-					'server is not in database, please contact portal support', undefined, client);
-				return false;
-			}
-
-			if (!command_options) {
-				message_reply(false, message.channel, message, message.author,
-					'could not get command option_2', guild_object, client);
-				return false;
-			}
-
-			if (command_options.premium && !guild_object.premium) {
-				message_reply(false, message.channel, message, message.author,
-					'server is not premium', guild_object, client);
-				return false;
-			}
-
-			if (command_options.auth && message.member) {
-				if (!is_authorised(guild_object, message.member)) {
-					message_reply(false, message.channel, message, message.author,
-						'you are not authorised to use this command', guild_object, client);
+				// is not a portal command
+				if (!is_portal_command) {
 					return false;
 				}
-			}
 
-			command_loader(message, cmd, args, type, command_options, path_to_command, guild_object);
+				if (!command_options) {
+					message_reply(false, message.channel, message, message.author,
+						'could not get command option', undefined, client);
+					return false;
+				}
+
+				if (!message) {
+					return false;
+				}
+
+				if (!message.guild) {
+					message_reply(false, message.channel, message, message.author,
+						'could not fetch guild of message', undefined, client);
+					return false;
+				}
+
+				fetch_guild(message.guild.id)
+					.then(guild_object => {
+						if (!guild_object) {
+							message_reply(false, message.channel, message, message.author,
+								'server is not in database, please contact portal support', undefined, client);
+							return false;
+						}
+
+						if (!command_options) {
+							message_reply(false, message.channel, message, message.author,
+								'could not get command option_2', guild_object, client);
+							return false;
+						}
+
+						if (command_options.premium && !guild_object.premium) {
+							message_reply(false, message.channel, message, message.author,
+								'server is not premium', guild_object, client);
+							return false;
+						}
+
+						if (command_options.auth && message.member) {
+							if (!is_authorised(guild_object, message.member)) {
+								message_reply(false, message.channel, message, message.author,
+									'you are not authorised to use this command', guild_object, client);
+								return false;
+							}
+						}
+
+						command_loader(message, cmd, args, type, command_options, path_to_command, guild_object);
+					})
+					.catch(error => {
+						console.log('could not fetch guild list' + error);
+						return false;
+					});
+			}
+			return;
 		})
-		.catch(error => {
-			console.log('could not fetch guild list' + error);
-			return false;
+		.catch(e => {
+			return;
 		});
 });
 
@@ -272,15 +290,10 @@ function command_loader(
 	if (type === 'none' && command_options.time === 0) {
 		require(`./commands/${path_to_command}/${cmd}.js`)(message, args, guild_object, client)
 			.then((response: ReturnPormise) => {
-				console.log('1. response :>> ', response);
-
-				console.log('command_p[tions.reply :>> ', command_options.reply);
 				if (response)
 					if (command_options.reply || (response.result === false && response.value !== ''))
 						message_reply(response.result, message.channel, message, message.author, response.value,
 							guild_object, client, command_options ? command_options.auto_delete : true);
-				// if (command_options.save_after)
-				// 	update_portal_managed_guilds(portal_managed_guilds_path, guild_list);
 			});
 		return true;
 	}
@@ -310,9 +323,8 @@ function command_loader(
 		return false;
 	}
 
-	require(`./commands/${path_to_command}/${cmd}.js`)(message, args, guild_object, client)
+	require(`./commands/${path_to_command}/${cmd}.js`)(message, args, guild_object, client, dispatchers)
 		.then((response: ReturnPormise) => {
-			console.log('response :>> ', response);
 			if (response) {
 				active_cooldowns[type === 'guild' ? 'guild' : 'member'].push({
 					member: message.author.id,
@@ -329,11 +341,6 @@ function command_loader(
 				}
 			}
 
-			// if (command_options.save_after)
-			// 	update_portal_managed_guilds(portal_managed_guilds_path, guild_list);
-
-			console.log(`${command_options} && ${response.value} && ${response.value} !== '' && (${command_options.reply} || ${response.result} === false) => `,
-				command_options && response.value && response.value !== '' && (command_options.reply || response.result === false));
 			if (command_options && response.value && response.value !== '' && (command_options.reply || response.result === false))
 				message_reply(response.result, message.channel, message, message.author,
 					response.value, guild_object, client, command_options.auto_delete);
@@ -347,7 +354,7 @@ function event_loader(event: string, args: any): void {
 	console.log(`├─ event-${event}`);
 	require(`./events/${event}.js`)(args)
 		.then((response: ReturnPormise) => {
-			if (event === 'messageReactionAdd' && response) {
+			if (event === 'messageReactionAdd' && response && response.result === true) {
 				const messageReaction = <MessageReaction>args.messageReaction;
 				if (messageReaction && messageReaction.message && messageReaction.message.guild) {
 					fetch_guild(messageReaction.message.guild.id)
@@ -356,7 +363,7 @@ function event_loader(event: string, args: any): void {
 								if (messageReaction.message.channel.id === guild_object.music_data.channel_id) {
 									const music_channel: TextChannel = args.messageReaction.message.guild.channels.cache
 										.find((channel: TextChannel) => channel.id === guild_object.music_data.channel_id);
-									// auto na trexei mono otan einai music reaction
+
 									music_channel
 										.send(`${args.user}, ${response.value}`)
 										.then(msg => { msg.delete({ timeout: 5000 }); })
@@ -386,10 +393,23 @@ async function portal_channel_handler(message: Message): Promise<boolean> {
 		if (!message.guild) return false;
 		fetch_guild(message.guild.id)
 			.then(guild_object => {
-				if (!guild_object) return true;
+				if (!guild_object)
+					return resolve(true);
 
 				if (included_in_url_list(message.channel.id, guild_object)) {
-					if (is_url(message.content)) {
+					if (message.content === './url') {
+						remove_url(guild_object.id, message.channel.id)
+							.then(r => {
+								message_reply(false, message.channel, message, message.author,
+									r ? 'successfully removed url channel' : 'failed to remove url channel',
+									guild_object, client);
+							})
+							.catch(e => {
+								message_reply(false, message.channel, message, message.author,
+									'failed to remove url channel', guild_object, client);
+							});
+					}
+					else if (is_url(message.content)) {
 						client_talk(client, guild_object, 'url');
 					}
 					else {
@@ -398,29 +418,52 @@ async function portal_channel_handler(message: Message): Promise<boolean> {
 							'url-only channel', guild_object, client);
 						message.delete();
 					}
+
 					return resolve(true);
 				}
 				else if (guild_object.music_data.channel_id === message.channel.id) {
-					start(client, message, message.content, guild_object, dispatchers)
-						.then(joined => {
-							// message_reply(
-							// 	joined.result, message.channel, message,
-							// 	message.author, joined.value, guild_list, client, true
-							// );
-							message.delete();
-						})
-						.catch(error => {
-							message_reply(
-								false, message.channel, message,
-								message.author, error, guild_object, client, true
-							);
-						});
+					if (message.content === './music') {
+						if (!message.guild) {
+							return resolve(true);
+						}
+
+						const dispatcher_object = dispatchers.find(d => d.id === guild_object.id)
+						const dispatcher = dispatcher_object ? dispatcher_object.dispatcher : undefined;
+
+						stop(guild_object, message.guild, dispatcher);
+						const music_data = new MusicData('null', 'null', []);
+						set_music_data(guild_object.id, music_data)
+							.then(r => {
+								message_reply(false, message.channel, message, message.author,
+									r ? 'successfully removed music channel' : 'failed to remove music channel',
+									guild_object, client);
+							})
+							.catch(e => {
+								message_reply(false, message.channel, message, message.author,
+									'failed to remove music channel', guild_object, client);
+							});
+					} else {
+						start(client, message, message.content, guild_object, dispatchers)
+							.then(joined => {
+								// message_reply(
+								// 	joined.result, message.channel, message,
+								// 	message.author, joined.value, guild_list, client, true
+								// );
+								message.delete();
+							})
+							.catch(error => {
+								message_reply(
+									false, message.channel, message,
+									message.author, error, guild_object, client, true
+								);
+							});
+					}
+
 					return resolve(true);
 				}
+
 				return resolve(false);
 			});
-
-		return resolve(false);
 	});
 }
